fix regression added in PR #143
While adding type-hinting the option to specify an input encoding
got ignored. This commit fixes this regression.

This commit also fixes the tests which call ansi2html as a command.
As the pytest documentation states, during test execution `stdin`
is set to a “null” object which will fail on attempts to read from it
because it is rarely desired to wait for interactive input when
running automated tests. So we also patch now `sys.stdin` using a
`io.TextIOWrapper` and wrapping any actual input in an `io.BytesIO`.

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>